Split functionality into lib.rs, with less in main.rs
