+++ kernel/src/lib.rs
@@ -1,0 +1,97 @@
+#![no_std]
+#![cfg_attr(test, no_main)]
+#![feature(custom_test_frameworks)]
+#![feature(abi_x86_interrupt)]
+#![test_runner(crate::test_runner)]
+#![reexport_test_harness_main = "test_main"]
+
+use core::panic::PanicInfo;
+
+pub mod interrupts;
+pub mod serial;
+pub mod vga_buffer;
+
+// init sets up critical core functions of the kernel.
+//
+pub fn init() {
+    interrupts::init();
+}
+
+// Testable represents a test function.
+//
+pub trait Testable {
+    fn run(&self) -> ();
+}
+
+// Wrap tests with debug statements.
+//
+impl<T> Testable for T
+where
+    T: Fn(),
+{
+    fn run(&self) {
+        serial_print!("{}...\t", core::any::type_name::<T>());
+        self();
+        serial_println!("[ok]");
+    }
+}
+
+// Entry point for the set of unit
+// tests.
+//
+pub fn test_runner(tests: &[&dyn Testable]) {
+    serial_println!("Running {} tests", tests.len());
+    for test in tests {
+        test.run();
+    }
+
+    exit_qemu(QemuExitCode::Success);
+}
+
+// Panic handler for tests.
+//
+pub fn test_panic_handler(info: &PanicInfo) -> ! {
+    serial_println!("[failed]\n");
+    serial_println!("Error: {}\n", info);
+    exit_qemu(QemuExitCode::Failed);
+    loop {}
+}
+
+// QemuExitCode represents the two valid
+// values for exiting QEMU.
+//
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[repr(u32)]
+pub enum QemuExitCode {
+    Success = 0x10,
+    Failed = 0x11,
+}
+
+// exit_qemu uses the 0xf4 I/O port to
+// instruct QEMU to exit with the given
+// exit code.
+//
+pub fn exit_qemu(exit_code: QemuExitCode) {
+    use x86_64::instructions::port::Port;
+
+    unsafe {
+        let mut port = Port::new(0xf4);
+        port.write(exit_code as u32);
+    }
+}
+
+// _start is the entry point for `cargo xtest`.
+//
+#[cfg(test)]
+#[no_mangle]
+pub extern "C" fn _start() -> ! {
+    init();
+    test_main();
+    loop {}
+}
+
+#[cfg(test)]
+#[panic_handler]
+fn panic(info: &PanicInfo) -> ! {
+    test_panic_handler(info)
+}
