+++ kernel/src/main.rs
@@ -2,14 +2,11 @@
 #![no_main]
 #![feature(abi_x86_interrupt)]
 #![feature(custom_test_frameworks)]
-#![test_runner(crate::test_runner)]
+#![test_runner(kernel::test_runner)]
 #![reexport_test_harness_main = "test_main"]
 
 use core::panic::PanicInfo;
-
-mod interrupts;
-mod serial;
-mod vga_buffer;
+use kernel::println;
 
 #[cfg(target_arch = "x86_64")]
 fn halt() {
@@ -31,13 +28,9 @@ fn panic(info: &PanicInfo) -> ! {
     }
 }
 
-fn init() {
-    interrupts::init();
-}
-
 #[no_mangle]
 pub extern "C" fn _start() -> ! {
-    init();
+    kernel::init();
 
     #[cfg(test)]
     test_main();
@@ -62,55 +55,5 @@ fn kmain() {
 #[cfg(test)]
 #[panic_handler]
 fn panic(info: &PanicInfo) -> ! {
-    serial_println!("[failed]\n");
-    serial_println!("Error: {}\n", info);
-    exit_qemu(QemuExitCode::Failed);
-    loop {
-        halt();
-    }
-}
-
-// Testable represents a test function.
-//
-pub trait Testable {
-    fn run(&self) -> ();
-}
-
-// Wrap tests with debug statements.
-//
-impl<T> Testable for T
-where
-    T: Fn(),
-{
-    fn run(&self) {
-        serial_print!("{}...\t", core::any::type_name::<T>());
-        self();
-        serial_println!("[ok]");
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(u32)]
-pub enum QemuExitCode {
-    Success = 0x10,
-    Failed = 0x11,
-}
-
-pub fn exit_qemu(exit_code: QemuExitCode) {
-    use x86_64::instructions::port::Port;
-
-    unsafe {
-        let mut port = Port::new(0xf4);
-        port.write(exit_code as u32);
-    }
-}
-
-#[cfg(test)]
-fn test_runner(tests: &[&dyn Testable]) {
-    serial_println!("Running {} tests", tests.len());
-    for test in tests {
-        test.run();
-    }
-
-    exit_qemu(QemuExitCode::Success);
+    kernel::test_panic_handler(info)
 }
